+++ Meraki.Api/Data/AdaptivePolicyPolicies.cs
@@ -15,14 +15,14 @@ public class AdaptivePolicyPolicies
 	/// </summary>
 	[ApiAccess(ApiAccess.ReadWrite)]
 	[DataMember(Name = "sourceGroup")]
-	public AdaptivePolicyPoliciesGroup? SourceGroup { get; set; }
+	public AdaptivePolicyPoliciesGroup SourceGroup { get; set; } = new();
 
 	/// <summary>
 	/// The destination adaptive policy group (requires one unique attribute)
 	/// </summary>
 	[ApiAccess(ApiAccess.ReadWrite)]
 	[DataMember(Name = "destinationGroup")]
-	public AdaptivePolicyPoliciesGroup? DestinationGroup { get; set; }
+	public AdaptivePolicyPoliciesGroup DestinationGroup { get; set; } = new();
 
 	/// <summary>
 	/// An ordered array of adaptive policy ACLs (each requires one unique attribute) that apply to this policy (default: [])
@@ -43,12 +43,12 @@ public class AdaptivePolicyPolicies
 	/// </summary>
 	[ApiAccess(ApiAccess.Read)]
 	[DataMember(Name = "createdAt")]
-	public string? CreatedAt { get; set; }
+	public DateTime CreatedAt { get; set; }
 
 	/// <summary>
 	/// Updated at
 	/// </summary>
 	[ApiAccess(ApiAccess.Read)]
 	[DataMember(Name = "updatedAt")]
-	public string? UpdatedAt { get; set; }
+	public DateTime UpdatedAt { get; set; }
 }